requirements.txt
Fabric==1.3.2
argparse==1.2.1
bitbucket-api==0.5.0
github3.py==1.0.0a1
cffi==0.8.6
chardet==2.0.1
command-not-found==0.2.44
cryptography==0.7.2
enum34==1.0.4
language-selector==0.1
medusa==0.5.4
meld3==0.6.5
ndg-httpsclient==0.3.3
paramiko==1.7.7.1
pbr==0.10.7
pyOpenSSL==0.14
pyasn1==0.1.7
pycparser==2.10
pycrypto==2.4.1
pycurl==7.19.0
requests==2.5.1
requests-oauthlib==0.4.2
sh==1.11
six==1.9.0
stevedore==1.2.0
supervisor==3.0a8
ufw==0.31.1-1
virtualenv==1.11.1
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2
wsgiref==0.1.2
